requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
